Refactor: created new dictionaries directory for each lang
