+++ src/config/i18n.config.ts
@@ -7,6 +7,7 @@ const locales = langs as unknown as string[]
 export const i18n: I18NConfig = {
   defaultLocale,
   locales,
+  localeDetection: false,
 }
 
 export type Locale = (typeof langs)[number]
