+++ src/dictionaries/default-dictionaries.ts
@@ -1,5 +1,5 @@
-import { enUS } from './default-language-collections/default-en-US'
-import { ptBR } from './default-language-collections/default-pt-BR'
+import { enUS } from './default-language-collections/default-en-US/default-en-US'
+import { ptBR } from './default-language-collections/default-pt-BR/default-pt-BR'
 
 export const defaultDictionary = {
   en: enUS,
