+++ src/dictionaries/default-language-collections/default-en-US/default-en-US.ts
@@ -1,0 +1,8 @@
+import { DefaultDictionary } from '../types/@dictionary.type'
+import { homePageEnUS } from './homePage-en-US'
+import { rdProjectPageEnUS } from './rdProjectPage-en-US'
+
+export const enUS: DefaultDictionary = {
+  home: homePageEnUS,
+  rdProjectPage: rdProjectPageEnUS,
+}
