+++ src/dictionaries/default-language-collections/default-pt-BR/default-pt-BR.ts
@@ -1,0 +1,8 @@
+import { DefaultDictionary } from '../types/@dictionary.type'
+import { homePagePtBR } from './homePage-pt-BR'
+import { rdProjectPagePtBR } from './rdProjectPage-pt-BR'
+
+export const ptBR: DefaultDictionary = {
+  home: homePagePtBR,
+  rdProjectPage: rdProjectPagePtBR,
+}
